gateway/requirements.txt
fastapi
uvicorn[standard]
httpx
python-jose[cryptography]
redis
prometheus-client
python-multipart
PyJWT
cryptography
requests
